extract utils to cfpq_data_devtools package
