+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup, find_packages
+from os.path import join, dirname
+
+setup(
+    name='cfpq_data_devtools',
+    version='1.0.0',
+    packages=find_packages(),
+    description='Development tools to work with data from CFPQ_Data repository',
+)
